feat(...) : 챗 가져오기 뷰모델로
